+++ include/box_party_pokemon_dropdown.h
@@ -11,12 +11,12 @@ enum
     BPPD_MOVE_INNER_Y
 };
 
-void AllocBoxPartyPokemonDropdowns(u8 num);
+void AllocBoxPartyPokemonDropdowns(u32 num);
 void FreeBoxPartyPokemonDropdowns(void);
-void SetBoxPartyPokemonDropdownMap2(u8 idx, u8 bgId, const void * src, u16 width, u16 height);
-void SetBoxPartyPokemonDropdownMap2Pos(u8 idx, u16 x, u16 y);
-void SetBoxPartyPokemonDropdownMap2Rect(u8 idx, u16 x, u16 y, u16 width, u16 height);
-void AdjustBoxPartyPokemonDropdownPos(u8 a0, u8 a1, s8 a2);
-void CopyBoxPartyPokemonDropdownToBgTilemapBuffer(u8 idx);
+void SetBoxPartyPokemonDropdownMap2(u32 idx, u32 bgId, const void * src, u16 width, u16 height);
+void SetBoxPartyPokemonDropdownMap2Pos(u32 idx, u16 x, u16 y);
+void SetBoxPartyPokemonDropdownMap2Rect(u32 idx, u16 x, u16 y, u16 width, u16 height);
+void AdjustBoxPartyPokemonDropdownPos(u32 a0, u32 a1, s8 a2);
+void CopyBoxPartyPokemonDropdownToBgTilemapBuffer(u32 idx);
 
 #endif //GUARD_BOX_PARTY_POKEMON_DROPDOWN_H
